Update feature fixed
Update settings feature was fixed to update the new settings within the app once the user pushed the Update button. Before, the feature updated the settings in the csv but did not read them back into the app. So, settings in the app were reverting to the old settings with a new file upload.